+++ src/Agents.jl
@@ -14,7 +14,7 @@ using Graphs
 using DataFrames
 using MacroTools
 using DynamicSumTypes
-export DynamicSumTypes
+export variant, variantof
 import ProgressMeter
 using Random
 using StaticArrays: SVector