+++ test/performance/multiagent_vs_union.jl
@@ -1,8 +1,8 @@
 # This file compares approaching multi-agent models in two ways:
 # 1) using different Types to represent different agents. This leads to type
-# stability within `step!`.
-# 2) using a single type with some extra property `type` or `kind`, and then do
-# an `if`-based branching on this type to dispatch to different functions.
+# instabilities.
+# 2) using @multiagent to enclose all types in a single one. This removes type
+# instabilities.
 
 # The result is that (2) is much faster.
 
@@ -45,6 +45,8 @@ end
     eight::Int64
 end
 
+const types = Union{GridAgentOne, GridAgentTwo, GridAgentThree, GridAgentFour, GridAgentFive, GridAgentSix}
+
 agent_step!(agent::GridAgentOne, model1) = randomwalk!(agent, model1)
 function agent_step!(agent::GridAgentTwo, model1)
     agent.one += rand(abmrng(model1))
@@ -60,13 +62,15 @@ function agent_step!(agent::GridAgentFour, model1)
     agent.one += sum(a.one for a in nearby_agents(agent, model1))
 end
 function agent_step!(agent::GridAgentFive, model1)
-    targets = filter!(a->a.one > 0.8, collect(nearby_agents(agent, model1, 3)))
-    idx = argmax(map(t->euclidean_distance(agent, t, model1), targets))
-    farthest = targets[idx]
-    walk!(agent, sign.(farthest.pos .- agent.pos), model1)
+    targets = filter!(a->a.one > 1.0, collect(types, nearby_agents(agent, model1, 3)))
+    if !isempty(targets)
+        idx = argmax(map(t->euclidean_distance(agent, t, model1), targets))
+        farthest = targets[idx]
+        walk!(agent, sign.(farthest.pos .- agent.pos), model1)
+    end
 end
 function agent_step!(agent::GridAgentSix, model1)
-    agent.eight += sum(rand(abmrng(model2), (0, 1)) for a in nearby_agents(agent, model1))
+    agent.eight += sum(rand(abmrng(model1), (0, 1)) for a in nearby_agents(agent, model1))
 end
 
 model1 = StandardABM(
@@ -89,67 +93,40 @@ end
 
 ################### DEFINITION 2 ###############
 
-@multiagent :opt_speed struct GridAgent2All(GridAgent{2})
-    @subagent struct GridAgent2One
-        one::Float64
-        two::Bool
-        three::Int
-    end
-    @subagent struct GridAgent2Two
-        one::Float64
-        two::Bool
-        four::Float64
-    end
-    @subagent struct GridAgent2Three
-        one::Float64
-        two::Bool
-        five::Bool
-    end
-    @subagent struct GridAgent2Four
-        one::Float64
-        two::Bool
-        six::Int16
-    end
-    @subagent struct GridAgent2Five
-        one::Float64
-        two::Bool
-        seven::Int32
-    end
-    @subagent struct GridAgent2Six
-        one::Float64
-        two::Bool
-        eight::Int64
-    end
-end
+agent_step!(agent, model2) = agent_step!(agent, model2, variant(agent))
 
-@dispatch agent_step!(agent::GridAgent2One, model2) = randomwalk!(agent, model2)
-@dispatch function agent_step!(agent::GridAgent2Two, model2)
+agent_step!(agent, model2, ::GridAgentOne) = randomwalk!(agent, model2)
+function agent_step!(agent, model2, ::GridAgentTwo)
     agent.one += rand(abmrng(model2))
     agent.two = rand(abmrng(model2), Bool)
 end
-@dispatch function agent_step!(agent::GridAgent2Three, model2)
-    if any(a-> kindof(a) == :gridagenttwo, nearby_agents(agent, model2))
+function agent_step!(agent, model2, ::GridAgentThree)
+    if any(a-> variant(a) isa GridAgentTwo, nearby_agents(agent, model2))
         agent.two = true
         randomwalk!(agent, model2)
     end
 end
-@dispatch function agent_step!(agent::GridAgent2Four, model2)
+function agent_step!(agent, model2, ::GridAgentFour)
     agent.one += sum(a.one for a in nearby_agents(agent, model2))
 end
-@dispatch function agent_step!(agent::GridAgent2Five, model2)
-    targets = filter!(a->a.one > 1.0, collect(nearby_agents(agent, model2, 3)))
+function agent_step!(agent, model2, ::GridAgentFive)
+    targets = filter!(a->a.one > 1.0, collect(GridAgentAll, nearby_agents(agent, model2, 3)))
     if !isempty(targets)
         idx = argmax(map(t->euclidean_distance(agent, t, model2), targets))
         farthest = targets[idx]
         walk!(agent, sign.(farthest.pos .- agent.pos), model2)
     end
 end
-@dispatch function agent_step!(agent::GridAgent2Six, model2)
+function agent_step!(agent, model2, ::GridAgentSix)
     agent.eight += sum(rand(abmrng(model2), (0, 1)) for a in nearby_agents(agent, model2))
 end
 
+@multiagent GridAgentAll(
+    GridAgentOne, GridAgentTwo, GridAgentThree, GridAgentFour, GridAgentFive, GridAgentSix
+) <: AbstractAgent
+
 model2 = StandardABM(
-    GridAgent2All,
+    GridAgentAll,
     GridSpace((15, 15));
     agent_step!,
     rng = MersenneTwister(42),
@@ -157,28 +134,37 @@ model2 = StandardABM(
 )
 
 for i in 1:500
-    add_agent!(GridAgent2One, model2, rand(abmrng(model2)), rand(abmrng(model2), Bool), i)
-    add_agent!(GridAgent2Two, model2, rand(abmrng(model2)), rand(abmrng(model2), Bool), Float64(i))
-    add_agent!(GridAgent2Three, model2, rand(abmrng(model2)), rand(abmrng(model2), Bool), true)
-    add_agent!(GridAgent2Four, model2, rand(abmrng(model2)), rand(abmrng(model2), Bool), Int16(i))
-    add_agent!(GridAgent2Five, model2, rand(abmrng(model2)), rand(abmrng(model2), Bool), Int32(i))
-    add_agent!(GridAgent2Six, model2, rand(abmrng(model2)), rand(abmrng(model2), Bool), i)
+    agent = GridAgentAll(GridAgentOne(model2, random_position(model2), rand(abmrng(model2)), rand(abmrng(model2), Bool), i))
+    add_agent_own_pos!(agent, model2)
+    agent = GridAgentAll(GridAgentTwo(model2, random_position(model2), rand(abmrng(model2)), rand(abmrng(model2), Bool), Float64(i)))
+    add_agent_own_pos!(agent, model2)
+    agent = GridAgentAll(GridAgentThree(model2, random_position(model2), rand(abmrng(model2)), rand(abmrng(model2), Bool), true))
+    add_agent_own_pos!(agent, model2)
+    agent = GridAgentAll(GridAgentFour(model2, random_position(model2), rand(abmrng(model2)), rand(abmrng(model2), Bool), Int16(i)))
+    add_agent_own_pos!(agent, model2)
+    agent = GridAgentAll(GridAgentFive(model2, random_position(model2), rand(abmrng(model2)), rand(abmrng(model2), Bool), Int32(i)))
+    add_agent_own_pos!(agent, model2)
+    agent = GridAgentAll(GridAgentSix(model2, random_position(model2), rand(abmrng(model2)), rand(abmrng(model2), Bool), i))
+    add_agent_own_pos!(agent, model2)
 end
 
 ################### Benchmarks ###############
 
-@btime step!($model1, 50)
-@btime step!($model2, 50) # repeat also with :opt_memory
+t1 = @belapsed step!($model1, 50)
+t2 = @belapsed step!($model2, 50)
 
 # Results:
-# multiple types: 3.732 s (39242250 allocations: 2.45 GiB)
-# @multiagent :opt_speed: 577.185 ms (25818000 allocations: 1.05 GiB)
-# @multiagent :opt_memory: 870.460 ms (25868000 allocations: 1.05 GiB)
+# multiple types: 1.849 s (34267900 allocations: 1.96 GiB)
+# @multiagent: 545.119 ms (22952850 allocations: 965.93 MiB)
 
-Base.summarysize(model1)
-Base.summarysize(model2) # repeat also with :opt_memory
+m1 = Base.summarysize(model1)
+m2 = Base.summarysize(model2)
 
 # Results:
-# multiple types: 491.20 KiB
-# @multiagent :opt_speed: 686.13 KiB
-# @multiagent :opt_memory: 563.12 KiB
+# multiple types: 543.496 KiB
+# @multiagent: 546.360 KiB
+
+println("Time to step the model with multiple types: $(t1) s")
+println("Time to step the model with @multiagent: $(t2) s")
+println("Memory occupied by the model with multiple types: $(m1/1000) Kib")
+println("Memory occupied by the model with @multiagent: $(m2/1000) Kib")